WIP towards a new event loop model
will run an event loop as a 'normal' thread, and poke events into a
var directly, then signal a semaphore.

eventually, want to have a single executable, and have a script call
something like (request-display 640@480) which will kick the script
into SDL, and start the event thread. otherwise the script will run
as is done in run_file