workaround for bug #1
appears to have been an off-by-one error. not happy with the solution
yet. was able to track it down to initial PUSHLIT stomping on the
current stack frame, and also repro'd it during normal operation of
bouncers test (without snapshotting involved). need to review this one
a little more carefully.